+++ src/files/integration.js
@@ -1,29 +1,28 @@
 import { registerFileAction, addNewFileMenuEntry, Entry, Permission, FileType } from "@nextcloud/files";
 import { t } from "@nextcloud/l10n";
 
-import { openInOverleafV6, createNewTexDocument } from "./overleaf_handlers.js";
+import { openInOverleaf, createNewTexDocument } from "./handlers.js";
+import { Overleaf_AppID } from "../overleaf.js";
 
-import OverleafV6Icon from "../../img/app-v6-bl.svg?raw";
-
-const OverleafV6_AppID = "overleafv6_nextcloud";
+import OverleafIcon from "../../img/app-v6-bl.svg?raw";
 
 /*** Open LaTeX Document Action ***/
 
 const openTexDocumentAction = {
 	id: "overleafv6-open",
-	displayName: () => t(OverleafV6_AppID, "Open in Overleaf V6"),
+	displayName: () => t(Overleaf_AppID, "Open in Overleaf V6"),
 	enabled: ({ nodes, view }) => { 
 		return nodes.length === 1
 			&& !nodes.some(({ permissions }) => (permissions & Permission.READ) === 0)
 			&& nodes.every(({ type }) => type === FileType.File)
 			&& nodes.every(({ mime }) => mime === "application/x-tex")
 	},
-	iconSvgInline: () => OverleafV6Icon,
+	iconSvgInline: () => OverleafIcon,
 	exec: async ({ nodes }) => {
-		await openInOverleafV6(nodes[0]);
+		await openInOverleaf(nodes[0]);
 	},
 	execBatch: async ({ nodes }) => {
-		await openInOverleafV6(nodes[0]);
+		await openInOverleaf(nodes[0]);
 		return nodes.map(_ => null);
 	},
 };
@@ -32,13 +31,15 @@ registerFileAction(openTexDocumentAction);
 
 /*** Create New LaTeX Document Action ***/
 
+/* Makes no sense right now, but could be useful in the future.
 const createNewTexDocumentAction = {
     id: "overleafv6-create",
-    displayName: t(OverleafV6_AppID, "Create new LaTeX document"),
-    iconSvgInline: OverleafV6Icon,
+    displayName: t(Overleaf_AppID, "Create new LaTeX document"),
+    iconSvgInline: OverleafIcon,
     handler: async (context, content) => {
         await createNewTexDocument(context, content);
     }
 };
 
 addNewFileMenuEntry(createNewTexDocumentAction);
+*/