+++ src/files/overleaf_handlers.js
@@ -1,7 +1,0 @@
-export async function openInOverleafV6(node) {
-	console.log("Opening file in Overleaf V6");
-}
-
-export async function createNewTexDocument(context, content) {
-    console.log("Creating new LaTeX document");
-}