+++ src/overleaf.js
@@ -1,0 +1,1 @@
+export const Overleaf_AppID = "overleafv6_nextcloud";